move to new file for better error msgs, create exp in test
